Minor code cleanup.

+++ lib/ruby_bugzilla.rb
@@ -33,7 +33,7 @@ class RubyBugzilla
       self.password ||= pw_from_file
     end
 
-    [self.username, self.password]
+    return self.username, self.password
   end
 
   # Ruby will catch and raise Erron::ENOENT: if there is no
@@ -41,22 +41,21 @@ class RubyBugzilla
   def self.credentials_from_file
     begin
       creds = YAML.load_file(CREDS_FILE)
-    rescue Errno::ENOENT => error
-      return [nil, nil]
+    rescue Errno::ENOENT
+      return nil, nil
     end
 
-    [creds[:bugzilla_credentials][:username],
-      creds[:bugzilla_credentials][:password]]
+    return creds[:bugzilla_credentials][:username], creds[:bugzilla_credentials][:password]
   end
 
   def self.options
     begin
       options = YAML.load_file(CREDS_FILE)
-    rescue Errno::ENOENT => error
-      return ["https://bugzilla.redhat.com/", false]
+    rescue Errno::ENOENT
+      return "https://bugzilla.redhat.com/", false
     end
-    [options[:bugzilla_options][:bugzilla_uri],
-      options[:bugzilla_options][:debug]]
+
+    return options[:bugzilla_options][:bugzilla_uri], options[:bugzilla_options][:debug]
   end
 
   def self.python_bugzilla_installed?
@@ -70,46 +69,39 @@ class RubyBugzilla
   end
 
   def self.clear_login!
-    if File.exists?(COOKIES_FILE) then
-      File.delete(COOKIES_FILE)
-    end
+    File.delete(COOKIES_FILE) if File.exists?(COOKIES_FILE)
   end
 
   def self.login!(username = nil, password = nil)
-    login_cmd = "#{CMD} "
-    output = "Already Logged In"
-    params = {}
-
     raise "Please install python-bugzilla" unless python_bugzilla_installed?
+    return "Already Logged In" if self.logged_in?
 
-    unless self.logged_in?
-      username, password = self.credentials(username, password)
-      uri_opt, debug_opt = self.options
+    username, password = self.credentials(username, password)
+    uri_opt, debug_opt = self.options
 
-      params["--bugzilla="] = "#{uri_opt}/xmlrpc.cgi" unless uri_opt.nil?
-      params["--debug"]     = nil if debug_opt
-      params["login"]       = [username, password]
+    params = {}
+    params["--bugzilla="] = "#{uri_opt}/xmlrpc.cgi" unless uri_opt.nil?
+    params["--debug"]     = nil if debug_opt
+    params["login"]       = [username, password]
 
-      begin
-        login_cmd_result = LinuxAdmin.run!(CMD, :params => params)
-      rescue => error
-        # A failed login attempt could result in a corrupt COOKIES_FILE
-        clear_login!
-        raise "#{self.string_command(CMD, params, password)} Failed.\n#{error}"
-      end
-      output = login_cmd_result.output
+    begin
+      login_cmd_result = LinuxAdmin.run!(CMD, :params => params)
+    rescue => error
+      # A failed login attempt could result in a corrupt COOKIES_FILE
+      clear_login!
+      raise "#{self.string_command(CMD, params, password)} Failed.\n#{error}"
     end
-    [self.string_command(CMD, params, password), output]
+
+    return self.string_command(CMD, params, password), login_cmd_result.output
   end
 
   def self.query(product, flag=nil, bug_status=nil, output_format=nil)
     raise "Please install python-bugzilla" unless python_bugzilla_installed?
     raise ArgumentError, "product cannot be nil" if product.nil?
 
-    uri_opt, debug_opt = self.options
+    uri_opt, _ = self.options
 
     params = {}
-
     params["--bugzilla="]     = "#{uri_opt}/xmlrpc.cgi" unless uri_opt.nil?
     params["query"]           = nil
     params["--product="]      = product
@@ -123,9 +115,8 @@ class RubyBugzilla
       raise "#{self.string_command(CMD, params)} Failed.\n#{error}"
     end
 
-    [self.string_command(CMD, params), query_cmd_result.output]
+    return self.string_command(CMD, params), query_cmd_result.output
   end
-
 
   #
   # Example Usage:
@@ -152,17 +143,17 @@ class RubyBugzilla
       raise ArgumentError, "bugids and options must be specified"
     end
 
-    uri_opt, debug_opt = self.options
-    params = {}
+    uri_opt, _ = self.options
 
-    params["--bugzilla="]     = "#{uri_opt}/xmlrpc.cgi" unless uri_opt.nil?
-    params["modify"]          = nil
+    params = {}
+    params["--bugzilla="] = "#{uri_opt}/xmlrpc.cgi" unless uri_opt.nil?
+    params["modify"]      = nil
 
     self.set_params_bugids(params, bugids)
     self.set_params_options(params, options)
 
     begin
-      modify_cmd_result = LinuxAdmin.run!(CMD, :params => params)
+      LinuxAdmin.run!(CMD, :params => params)
     rescue => error
       raise "#{self.string_command(CMD, params)} Failed.\n#{error}"
     end
